local-mcp-orchestrator/requirements.txt
llama-cpp-python
langchain
langchain-community
prompt_toolkit
PyYAML
duckduckgo-search
requests
mcp
